requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-chroma
faiss-cpu
pypdf
python-dotenv
